Join small classes into one file
