--- exporters/decompressors/base_decompressor.py
@@ -1,6 +1,0 @@
-from exporters.pipeline.base_pipeline_item import BasePipelineItem
-
-
-class BaseDecompressor(BasePipelineItem):
-    def decompress(self):
-        raise NotImplementedError()
