--- exporters/deserializers/base_deserializer.py
@@ -1,6 +1,0 @@
-from exporters.pipeline.base_pipeline_item import BasePipelineItem
-
-
-class BaseDeserializer(BasePipelineItem):
-    def deserialize(self, stream):
-        raise NotImplementedError()
