+++ exporters/module_loader.py
@@ -21,11 +21,11 @@ class ModuleLoader(object):
         return self._load_module(options, metadata, BaseWriter, **kwargs)
 
     def load_decompressor(self, options, metadata, **kwargs):
-        from exporters.decompressors.base_decompressor import BaseDecompressor
+        from exporters.decompressors import BaseDecompressor
         return self._load_module(options, metadata, BaseDecompressor, **kwargs)
 
     def load_deserializer(self, options, metadata, **kwargs):
-        from exporters.deserializers.base_deserializer import BaseDeserializer
+        from exporters.deserializers import BaseDeserializer
         return self._load_module(options, metadata, BaseDeserializer, **kwargs)
 
     def load_persistence(self, options, metadata, **kwargs):
