+++ tests/test_exporter_config.py
@@ -245,10 +245,10 @@ class ConfigValidationTest(unittest.TestCase):
     def test_stream_only_sections(self):
         config = valid_config_with_updates({
             "decompressor": {
-                "name": "exporters.decompressors.zlib_decompressor.ZLibDecompressor",
+                "name": "exporters.decompressors.ZLibDecompressor",
             },
             "deserializer": {
-                "name": "exporters.deserializers.csv_deserializer.CSVDeserializer",
+                "name": "exporters.deserializers.CSVDeserializer",
             },
         })
         with self.assertRaises(ConfigurationError) as cm:
@@ -268,10 +268,10 @@ class ConfigValidationTest(unittest.TestCase):
                 }
             },
             "decompressor": {
-                "name": "exporters.decompressors.zlib_decompressor.ZLibDecompressor",
+                "name": "exporters.decompressors.ZLibDecompressor",
             },
             "deserializer": {
-                "name": "exporters.deserializers.csv_deserializer.CSVDeserializer",
+                "name": "exporters.deserializers.CSVDeserializer",
             },
         })
         check_for_errors(config)  # should not raise
